moved test code to use new factory methods getClusterInfo, getClusterRepository.
