+++ lib/comoonics/cluster/test/test_ClusterInfo.py
@@ -1,6 +1,6 @@
 from BaseClusterTestClass import baseClusterTestClass
-from comoonics.cluster.ComClusterRepository import ClusterMacNotFoundException
 from comoonics import ComLog
+from comoonics.cluster import getClusterRepository, getClusterInfo
 
 import unittest
 
@@ -10,17 +10,15 @@ class test_ClusterInfo(baseClusterTestClass):
     """   
     def init(self):
         import os.path
-        from comoonics.cluster.ComClusterRepository import ClusterRepository
-        from comoonics.cluster.ComClusterInfo import ClusterInfo
         import logging
         from comoonics import ComSystem
         ComSystem.setExecMode(ComSystem.SIMULATE)
         super(test_ClusterInfo, self).init()
         #create comclusterRepository Object
-        self.clusterRepository = ClusterRepository(os.path.join(self._testpath, "cluster2.conf"))
+        self.clusterRepository = getClusterRepository(os.path.join(self._testpath, "cluster2.conf"))
 
         #create comclusterinfo object
-        self.clusterInfo = ClusterInfo(self.clusterRepository)  
+        self.clusterInfo = getClusterInfo(self.clusterRepository)  
 
         # setup the cashes for clustat for redhat cluster
         ComLog.setLevel(logging.DEBUG)
