+++ lib/comoonics/cluster/test/test_ClusterRepository.py
@@ -1,5 +1,5 @@
 from BaseClusterTestClass import baseClusterTestClass
-from comoonics.cluster.ComClusterRepository import ClusterMacNotFoundException
+from comoonics.cluster import getClusterRepository, ClusterMacNotFoundException
 
 import unittest
 
@@ -9,10 +9,9 @@ class test_ClusterRepository(baseClusterTestClass):
     """                    
     def setUp(self):
         import os.path
-        from comoonics.cluster.ComClusterRepository import ClusterRepository
         super(test_ClusterRepository, self).setUp()
         #create comclusterRepository Object
-        self.clusterRepository = ClusterRepository(os.path.join(self._testpath, "cluster2.conf"))
+        self.clusterRepository = getClusterRepository(os.path.join(self._testpath, "cluster2.conf"))
 
     def testGetnodename(self):
         # except first nic because it does not have an mac-address
@@ -27,11 +26,37 @@ class test_ClusterRepository(baseClusterTestClass):
         for i in range(len(_tmp)):
             self.assertEqual(_tmp[i]["nodeid"], str(self.clusterRepository.getNodeId(_tmp[i]["mac"])))
         self.assertRaises(ClusterMacNotFoundException, self.clusterRepository.getNodeId,"murks")
+        
+    def testStr(self):
+        self.assertEquals(str(self.clusterRepository), "ComoonicsClusterRepository(nodes: 3)")
+        
+    def testClusterInfoClass(self):
+        from comoonics.cluster.ComClusterInfo import ComoonicsClusterInfo
+        self.assertEquals(self.clusterRepository.getClusterInfoClass(), ComoonicsClusterInfo)
+
+class test_ClusterRepository2(unittest.TestCase):      
+    """
+    Methods from RedhatClusterRepository
+    """                    
+    def setUp(self):
+        self.clusterRepository = getClusterRepository()
+    def testCreateEmpty(self):
+        from comoonics.cluster.ComClusterRepository import ClusterRepository
+        #create comclusterRepository Object
+        self.assertTrue(isinstance(self.clusterRepository, ClusterRepository))
+        
+    def testStr(self):
+        self.assertEquals(str(self.clusterRepository), "ClusterRepository(nodes: 0)")
+
+    def testClusterInfoClass(self):
+        from comoonics.cluster.ComClusterInfo import ClusterInfo
+        self.assertEquals(self.clusterRepository.getClusterInfoClass(), ClusterInfo)
 
 def test_main():
     try:
         from test import test_support
         test_support.run_unittest(test_ClusterRepository)
+        test_support.run_unittest(test_ClusterRepository2)
     except ImportError:
         unittest.main()
